refactor: modularize complexity checker
