+++ tools/quality/code/check-code-complexity/constants.ts
@@ -1,0 +1,6 @@
+export const MAX_CYCLOMATIC_COMPLEXITY = 10
+export const MAX_COGNITIVE_COMPLEXITY = 15
+export const MAX_NESTING = 4
+
+export const DEFAULT_SRC_DIR = 'src'
+export const IGNORED_DIRECTORIES = ['node_modules', '.next', 'dist', 'build']
