+++ tools/quality/code/check-code-complexity/function-pattern.ts
@@ -1,0 +1,1 @@
+export const FUNCTION_PATTERN = /(?:function|const|async\s+(?:function)?|export\s+(?:async\s+)?function)\s+(\w+)/g
