+++ tools/quality/code/check-code-complexity/run-check-code-complexity.ts
@@ -1,0 +1,8 @@
+import { analyzeComplexity } from './walk-directory'
+import { buildSummary } from './build-summary'
+
+export const runCheckCodeComplexity = (): string => {
+  const results = analyzeComplexity()
+  const summary = buildSummary(results)
+  return JSON.stringify(summary, null, 2)
+}
